deployment/requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
plotly
scikit.learn==1.3.0
xgboost
pillow